revert: undo list_app.rs directory module split
Reverts f3a1519 and 8c36ba2. The split into mod.rs/actions.rs/render.rs
scattered tightly-coupled TUI state machine logic across 3 files, making
it harder to trace features end-to-end. The shared TUI pattern
extractions (keybindings, modals) are also reverted and will be
re-applied on the single file in a follow-up commit.
